Implementing Scroll-to-load-content on search page.
A search conducted on the form in header or on search page now
stores the entry-ids of matching entries in an HTTP session.
The search twig template displays only first 10 entries. As the
user scrolls down and reaches near-bottom, the next 10 entries
are retrieved from the entry-ids stored in HTTP session and
appended to DOM.

A new craft plugin was created to set/get session variables.
See: craft/plugins/twigsession
To retrieve a session variable's value, simply use the following
in a Twig template:
craft.twigSession.get('variable_name')
To set a session variable, use the following:
craft.twigSession.set('variable_name', 'variable_value')

In the search handler, we save a session variable named
`searchEntryIds` with value as an array of ids of entries
matching search term.

An API route `/search_custom/entries` was created that accepts
`offset` and `limit` as two variables, and based on them
splices the stored session variable value of `searchEntryIds`
and returns the markup for the spliced entries.

A javascript scroll handler was created in `search.js` to invoke
the above API and append the results when user scrolls to the
bottom.

In addition, search functionality was also optimized as follows:
 - duplicate entries are removed in search results.
 - remove re-ordering of matching entries in Twig template as they
   are already ordered in `MpSearchService.php#search`.