fixing the regression where letters section on contributor profile page will not show Load More. Also, addressing the corner case where Load More will show up even though no entries exist if exactly 10 entries (or multiple of 10 entries)  exist under a section.
the backend API to fetch contributor entries based on offset and
limit now returns 'count' and 'total_count' which are used by
javaqscript to determine if Load More should be hidden or shown
and if the entire section should be hidden (total_count = 0). 'count'
returns the number of entries that were fetched for the given offset
and limit.