Hitting Close on My Images/Documents should redirect to current page
Enhancement request:
From the HOME PAGE, click on MY ACCOUNT, MY IMAGES or MY DOCUMENTS.
When a user is finished adding images or documents and clicks on “CLOSE” they are redirected to the HOME PAGE. Change it so that they are still on / redirected to whatever page they were on when they opened the MY IMAGES or the MY DOCUMENTS widget.

Fix:
Instad of using document.location.href='/', use window.history.back() to navigate to the page from where user clicke My Images or My Documents.

Testing:
Tested on Chrome, Safari & Firefox on Desktop.

Other Info:
Task List MARIN POST 04-20-19
2. LANDING PAGE after MY IMAGES or MY DOCUMENTS
From the HOME PAGE, click on MY ACCOUNT, MY IMAGES or MY DOCUMENTS.
When a user is finished adding images or documents and clicks on “CLOSE” they are redirected to the HOME PAGE. Change it so that they are still on / redirected to whatever page they were on when they opened the MY IMAGES or the MY DOCUMENTS widget.