Display count of search results on search page.
If the number of search results is exactly one, a phrase displays
to the right of radio buttons aligned right "1 result found".

If the number of search results is more than one, a phrase displays
to the right of the radio buttons.

It was tested on mobile and desktop.

Important: As part of this commit, a regression bug was identified
and fixed. The bug is that some search results that should match
the search do not appear. Read on to understand the fix for it.

The auto loading algorithm determines the offset by counting the
number of posts currently displayed on the search post. Currently,
it counts by using $('posts.search li') expression. This counts
li elements that are nested within parent li elements as well. The
fix was to change the expression to $('posts.search > li') to count
only top level li elements each that uniquely determines a search result.
See search.js changes.