Adding "Post a Blog | a Notice | a Letter | News | Media" link under My Accounts -> My Blog/My Notices/My News/My Media/My Letters pages respectively.
<h2> housing the title of these pages previously is now inlined with
a link (<a>) element as children of a containing div. CSS was adjusted
so that the title and link text are near baseline aligned and the rest
of spacings are preserved. During the synopsis, an undefined (dead) css class
'comment-count' was found and removed from <h2> inside account/_layout.html.

This commit pertains to Change Request # 5 in Task list "MARIN POST
04-20-19".

5. Add “CREATE A …(POST)” links on each page under MY ACCOUNT / MY BLOGS, MY
NOTICES, etc.
Some users have asked if we can add a “Create a Blog” or a “Create a Notice” and so forth, link
to the pages under MY ACCOUNT for MY BLOGS, MY NOTICES, MY NEWS, MY MEDIA and MY
LETTERS…. As a quick way to get to the “create” form pages for each.
This would be located to the upper right hand side of the center div, as shown below. It is
aligned with the page title, “My Blog Posts” but registered to the right.
It will be important to test this UI change when it is viewed in the narrowest view for a mobile
phone.