Fixing Bold, Underline & Italics on admin+public redactor editor.
This is a known redactor bug fixed in Redactor 3.0. However, we are
stuck with Redactor 2.0 because of Craft CMS 2.0 compatibility issues.

Therefore, the fix is to manually port the recommended fix into the
plugin code.

See the following StackOverflow article describing the issue and fix:
https://stackoverflow.com/questions/43674400/redactor-editor-text-format-issues-with-chrome-version-58

See the GitHub issue page for the known issue:
https://github.com/concrete5/concrete5/issues/5406