Support embedding an image in blog/notice content.
Users can now embed an image in blog/notice content. An 'Add Image'
button shows up to the right of the underline button in the redactor
editor on the create/edit blog/notice pages. User taps 'Add Image',
image picker shows up that behaves exactly the way the existing image
picker shows up. It displays the user's media library and gives him/her
an option to upload, delete, select an image or cancel the dialog.
Upon choosing an image and hitting OK, the selected image is embedded
into the editor. User can further edit the image by clicking on the
embedded image and hitting the 'Edit' button atop it. For example: the
image can be given a title, can be made into a link and the link can
be opened in a new tab, can be aligned (left/center/right), can be
deleted, etc.

A custom Redactor plugin - 'addimage' - adds the 'Add Image' button to
UI, shows image picker when this button is tapped, and embeds the selected
image upon hitting ok. The file public/js/redactor/addimage.js
encapsulates all logic of this plugin.

The blog and notice twig templates now includes a new twig template
called as 'image-redactor'. This twig template encapsulates the
second image picker. This image picker is separate to the one used
by 'Image' field of the form and intentionally kept separate so as
to not interfere with the existing picker's functionality. See
craft/templates/_form/image-redactor.html for more details.

Due to the existence of two pickers, there is logic in place to sync
the images in both pickers if they are added or deleted in either of
the two pickers.

This change request pertains to #9 in the 06-16-19 task list below:

"
9. ADD NEW TEXT EDITOR “INSERT IMAGE” FEATURE FOR USERS
Our current ADMIN View offers the ability to “Insert Image” into the content of a BLOG or NOTICE. I would like offer this feature from our users, so they can upload and place images on their blog / notice pages in between text content. The image and text will always be separated (no wrapped text).
However, I want the “insert image” to bring up their personal user Image Library popup. So they can load it into their Library and then insert it with a double click.
This raises issues about linking the feature to AWS, as we discussed.
TASK: Add the capability for users to add images to their post content.
NOTE: This capability already exists in the Redactor text editor in the ADMIN view except it loads the image into the main Admin library on Digital Ocean.
"