implementing ability to preview documents in document browser
Under My Accounts->My Documents, Document Browser shows PDFs.
Currently, user can only delete a PDF but can't preview it.
With this change, user can double click a PDF and preview it
in a new tab.

An attribute data-asset-url was added to the list-item representing
a PDF document (asset in craft lingo). The value of this attribute
is populated to be the URL (using .url). A double click listener
is implemented that first marks the PDF as selected and then gets
the data-asset-url value from it and uses the window.open(<url>,
_blank) method to open the URL in a new tab.

Tested on Chrome with scenarios:
1) Double Click an existing PDF
2) Upload a new document and double click an existing PDF
3) Upload a new document and double click the uploaded PDF.