Open pasted links in redactor editor in new tab
Bob discovered a bug where if one pastes text containing links
in redactor editor, those links would open in same tab. The
desired behaviour is to open every link in new tab.

The fix is to add target=_blank to every link (a) element in
the generated HTML of the text pasted in the 'paste' #insert
callback.