Implementing embedding high res images in posts.
These changes enable users to upload an image and embed it in their
posts via the redactor editor *without* breaking the aspect ratio
or the resolution of the image. User now click on the 'Insert Image'
button in the post editor to invoke the image picker. The image picker
allows them to upload an image. Upon uploading the image, a thumbnail
depicting the image appears in the image picker. Note - we got rid of
the image titles per Bob's request. The text above the gallery is a
global message "my/embedded/images". Upon selecting an uploaded image
and hitting OK OR double clicking an image, embeds the originally
uploaded image into the editor that is scaled to fit, i.e, it will
resize (to lesser size) if its original width exceeded the available
editor width OR it will retain its original dimensions if the width
of the editor is more than its original width. Like before, users
can edit the embedded image, as in, clicking it and hitting Edit
button. A link or alt-text may be added.

Language was also changed a bit.
Add an Image -> Add Post Image
Replace Image -> Replace Post Image
Add Image (in editor) -> Insert Image

A new "Asset Source" of kind images was created in Admin called
"Embedded Images" (sourceId = 4). An Image Transform "Embedded
Image Thumb" was created for creating a "scaled to fit" thumbnail
with width 190px to show the uploaded image in the image picker.
A global message "my/embedded/images" was defined to show in the
image picker for this use case. The images are uploaded in the
emimg subfolder of the bucket. These have been done in DEV ADMIN
after testing on LOCAL. But NOT in PROD yet pending Review.

One change made in the S3 Direct plugin is to return the
originalUrl in addition to the transformed URL in the response
of the updateAssetsIndex API endpoint. The originalUrl is defined
for the embedded image whereas the url (transformed) is used to
display the uploaded image in picker.
See: craft/plugins/s3direct/controllers/S3DirectController.php

CSS changes were done to create a 5px padding around the image
thumbnails so that selection (gray background) is visible now
that we don't have the titles below the images for this image
picker.

Now that we have the embedded images in their own folder, the logic
to delete images in "embedded" image picker when deleted in "post"
image picker and viceversa has been yanked.
See: git diff craft/templates/_form/image-redactor.html
& git diff craft/templates/_form/image.html

Ofcourse the 'Insert Image' button in the public redactor editor
was re-instated.